lib: skip source maps in node_modules
Skipping source maps in `node_modules` by default improves the general
performance. Add `module.setSourceMapsSupport(enabled, options)` to
enable source  maps in `node_modules` if it is needed. This moves
all source maps related API to `node:module` and this a step to
promote the source maps API to stable.

Files in `node_modules` are not authored by the user directly and the
original sources are less relevant to the user.